==> 1.1_households_basic_data_decod.ipynb <==
import pandas as pd

try:
    home_basic = pd.read_csv('home_basic.csv')

except FileNotFoundError:
    url = 'https://drive.google.com/file/d/1M6UE6y24_uCjW8flv3tl4uT8Ge7s7mxI/view?usp=sharing'
    path = 'https://drive.google.com/uc?id=' + url.split('/')[-2]
    home_basic = pd.read_csv(path)

home_basic.head()

home_basic.rename(columns={'DB010': 'Year of survey'}, inplace=True)

home_basic.head()

home_basic.rename(columns={'DB020': 'Country'}, inplace=True)

home_basic.head()

home_basic.rename(columns={'DB030': 'Household ID'}, inplace=True)

home_basic.head()

home_basic.rename(columns={'DB040': 'Region'}, inplace=True)

mapped_region = {'ES11': 'Galicia', 
                 'ES12': 'Asturias', 
                 'ES13': 'Cantabria', 
                 'ES21': 'Basque Country', 
                 'ES22': 'Navarre',
                 'ES23': 'La Rioja',
                 'ES24': 'Aragon',
                 'ES30': 'Comunity of Madrid',
                 'ES41': 'Castile and Leon',
                 'ES42': 'Castile–La Mancha',
                 'ES43': 'Extremadura',
                 'ES51': 'Catalonia',
                 'ES52': 'Valencian Community',
                 'ES53': 'Balearic Islands',
                 'ES61': 'Andalusia',
                 'ES62': 'Region of Murcia',
                 'ES63': 'Ceuta',
                 'ES64': 'Melilla',
                 'ES70': 'Canary Islands',
                 'ESZZ': 'Extra-regional'}

home_basic['Region'] = home_basic['Region'].map(mapped_region)

home_basic.head()

home_basic.rename(columns={'DB040_F': 'Region (flag)'}, inplace=True)

mapped_region_flag = {-1: 'Missing', 1: 'Filled'}

home_basic['Region (flag)'] = home_basic['Region (flag)'].map(mapped_region_flag)

home_basic.head()

home_basic.rename(columns={'DB060': 'Primary sampling units (PSU)'}, inplace=True)

home_basic.head()

home_basic.rename(columns={'DB060_F': 'Primary sampling units (PSU) (flag)'}, inplace=True)

mapped_psu_flag = {1: 'Rotation is implemented at PSU level (the PSU rotates in and out of the sample)', 
                   2: 'Rotation is implemented at SSU or household level (The PSU remains in the sample for the entire duration of EU-SILC)', 
                  -2: 'Not applicable (no first or second sampling stage)'}

home_basic['Primary sampling units (PSU) (flag)'] = home_basic['Primary sampling units (PSU) (flag)'].map(mapped_psu_flag)

home_basic.head()

home_basic.rename(columns={'DB090': 'Houshold cross-sectional weight'}, inplace=True)

home_basic.head()

home_basic.rename(columns={'DB090_F': 'Houshold cross-sectional weight (flag)'}, inplace=True)

home_basic['Houshold cross-sectional weight (flag)'] = home_basic['Houshold cross-sectional weight (flag)'].map({1: 'Filled'})

home_basic.head()

home_basic.rename(columns={'DB100': 'Degree of urbanisation'}, inplace=True)

mapped_density = {1: 'Densely-populated area', 2: 'Intermediate area', 3: 'Thinly-populated area'}

home_basic['Degree of urbanisation'] = home_basic['Degree of urbanisation'].map(mapped_density)

home_basic.head()

home_basic.rename(columns={'DB100_F': 'Degree of urbanisation (flag)'}, inplace=True)

home_basic['Degree of urbanisation (flag)'] = home_basic['Degree of urbanisation (flag)'].map({1: 'Filled'})

home_basic.head()

for c in home_basic.columns:
    print(c)

home_basic.to_csv('home_basic_decod.csv')
