==> 4_feature_selection.ipynb <==
from ipynb.fs.full.my_functions import *
import pandas as pd
import numpy as np
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import OneHotEncoder
from sklearn.feature_selection import SelectKBest
from sklearn.feature_selection import f_classif
from sklearn.feature_selection import chi2

try:
    data = pd.read_csv('variables.csv')

except FileNotFoundError:
    url = 'https://raw.githubusercontent.com/deividvalerius/Multidimensional-Poverty-Predictor/master/Data/variables.csv'
    data = pd.read_csv(url)

data.info()

#First I scale the numerical variables. I use MinMaxScaler. Since I intend to enconde my categorical variables this way all
#my variables will have values between 0 and 1.

num_var = data[['age', 'years_worked', 'hours_week_worked', 'adjusted_income', 'proportion_social_welfare']]

scaler = MinMaxScaler()

scaler.fit(num_var)
scaled_var = pd.DataFrame(scaler.transform(num_var), columns=num_var.columns)
scaled_var.head()

#I weight the DataFrame before applying the hipothesis tests.

#weighted_df from my_functions takes a dataframe and returns a weighted version of the inputed dataframe.

scaled_var['weight'] = data.weight
scaled_var['material_deprivation'] = data.material_deprivation.map({'Yes': 1, 'No': 0})

weighted_scaled_var = weighted_df(scaled_var)

#I perform the analysis of variance

#weighted_cat from my_functions takes a categorical column and returns a weighted version of the inputed column.

X = weighted_scaled_var[weighted_scaled_var.columns[:-1]]
y = weighted_scaled_var.material_deprivation
                                                                 
fs = SelectKBest(score_func=f_classif, k='all').fit(X, y)

#I want to know which of my scaled variables got better pvalue scores.

#This dictionary has the pvalues as keys and the variable names as values.
scaled_pvalues = {}
for i in range(len(fs.pvalues_)):
    scaled_pvalues[fs.pvalues_[i]] = scaled_var.columns[i]

#Having the pvalues as keys makes sorting them easy. Below a list of all pvalues from smallest to largest.
for i in range(len(sorted(scaled_pvalues))):
    print('{}: {}'.format(scaled_pvalues[sorted(scaled_pvalues)[i]], sorted(scaled_pvalues)[i]))

#I weigth my categorical columns.

cat_var = data[['sex', 
                'civil_status', 
                'familial_status', 
                'region', 
                'population_density', 
                'citizenship', 
                'tenure_status',
                'education_level',
                'working_status', 
                'occupation',
                'bad_health']].reset_index(drop=True)

cat_var['weight'] = data.weight
cat_var['material_deprivation'] = data.material_deprivation.map({'Yes': 1, 'No': 0})

weighted_cat_var = weighted_df(cat_var)

#Saving and droping my independent variable after weighting. Do not need it in the dataframe for the encoding.

weighted_material_deprivation = weighted_cat_var.material_deprivation.copy()
weighted_cat_var = weighted_cat_var.drop(['material_deprivation'], axis=1)

#Defining the categories for the OneHotEncoder categories parameter.

categories = []
for i in range(len(weighted_cat_var.columns)):
    categories.append([])
    for value in weighted_cat_var[weighted_cat_var.columns[i]].unique():
        categories[i].append(value)

for i in range(len(categories)):
    print(categories[i])

#Encoding the dataframe

ohe = OneHotEncoder(categories=categories)
ohe.fit(weighted_cat_var)
onehot_var = ohe.transform(weighted_cat_var).toarray()

onehot_var

#Performing the chi square test.

X = onehot_var
y = weighted_material_deprivation

ch2 = SelectKBest(score_func=chi2, k='all').fit(X, y)

#Creating new names for the encoded columns

onehot_var_names = []

for i in range(len(categories)):
    for value in categories[i]:
        onehot_var_names.append(weighted_cat_var.columns[i] + '_' + value.lower().replace(' ', '_'))
        
onehot_var_names[:5]

#Just as before I want to know which variables got better pvalue score.

#Dictionary with the pvalues as keys and the variable names as values.
one_hot_pvalues = {}
for i in range(len(ch2.pvalues_)):
    one_hot_pvalues[ch2.pvalues_[i]] = onehot_var_names[i]
    

#This dictionary filters out variables with less than 99.9% significance. 
one_hot_pvalues_filtered = {}
for pvalue in one_hot_pvalues:
    if pvalue < 0.001:
        one_hot_pvalues_filtered[pvalue] = one_hot_pvalues[pvalue]

#List of all filtered pvalues from smallest to largest.
for i in range(len(sorted(one_hot_pvalues_filtered))):
    print('{}: {}'.format(one_hot_pvalues_filtered[sorted(one_hot_pvalues_filtered)[i]], sorted(one_hot_pvalues_filtered)[i]))

#Resulting dataframe after weighting and feature engeeniring.

one_hot_df = pd.DataFrame({onehot_var_names[i]: onehot_var.transpose()[i] for i in range(len(onehot_var_names))})

all_features = pd.concat([one_hot_df, weighted_scaled_var], axis=1)

all_features.head()

#Selected feature list

selected_cat_features = [one_hot_pvalues_filtered[key] for key in one_hot_pvalues_filtered]
selected_num_features = [scaled_pvalues[key] for key in scaled_pvalues]
selected_features = selected_cat_features + selected_num_features
selected_features.append('material_deprivation')

selected_features_df = all_features[selected_features]
selected_features_df.head()

#Exporting final table for modeling

selected_features_df.to_csv('to_model.csv', index=False)
