==> 5.1_logistic_regression.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from imblearn.under_sampling import RandomUnderSampler
from imblearn.over_sampling import RandomOverSampler
from imblearn.over_sampling import SMOTE
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report

try:
    data = pd.read_csv('to_model.csv')

except FileNotFoundError:
    url = 'https://raw.githubusercontent.com/deividvalerius/Multidimensional-Poverty-Predictor/master/Data/to_model.csv'
    data = pd.read_csv(url)

data.head()

X = data.drop(['material_deprivation'], axis=1)
y = data.material_deprivation

X_train, X_test, y_train, y_test = train_test_split(X, y)

#Imbalanced data

logreg = LogisticRegression(max_iter=1000, solver='lbfgs')
logreg.fit(X_train, y_train)
y_pred = logreg.predict(X_test)
print(classification_report(y_test, y_pred))

#Balanced data - Undersample

undersampler = RandomUnderSampler(sampling_strategy='majority')
X_train_us, y_train_us = undersampler.fit_resample(X_train, y_train)

logreg = LogisticRegression(max_iter=1000)

logreg.fit(X_train_us, y_train_us)
y_pred = logreg.predict(X_test)
print(classification_report(y_test, y_pred))

#Balanced data - Oversample

oversampler = RandomOverSampler(sampling_strategy='minority')
X_train_os, y_train_os = oversampler.fit_resample(X_train, y_train)

logreg = LogisticRegression(max_iter=1000)

logreg.fit(X_train_os, y_train_os)
y_pred = logreg.predict(X_test)
print(classification_report(y_test, y_pred))

#Balanced data - SMOTE

smote = SMOTE()
X_train_smote, y_train_smote = smote.fit_resample(X_train, y_train)

logreg = LogisticRegression(max_iter=1000)

logreg.fit(X_train_smote, y_train_smote)
y_pred = logreg.predict(X_test)
print(classification_report(y_test, y_pred))
